+++ projects/ngneat/hot-toast/src/lib/components/animated-icon/animated-icon.component.ts
@@ -5,6 +5,7 @@ import { IconTheme } from '../../hot-toast.model';
   selector: 'hot-toast-animated-icon',
   templateUrl: './animated-icon.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
 })
 export class AnimatedIconComponent {
   @Input() iconTheme: IconTheme;