+++ projects/ngneat/hot-toast/src/lib/components/indicator/icons/info/info.component.ts
@@ -5,6 +5,7 @@ import { IconTheme } from '../../../../hot-toast.model';
   selector: 'hot-toast-info',
   templateUrl: './info.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
 })
 export class InfoComponent {
   @Input() theme: IconTheme;